+++ src/main/java/twilightforest/entity/ai/HoverThenDropGoal.java
@@ -95,7 +95,7 @@ public class HoverThenDropGoal extends HoverBaseGoal<SnowQueenEntity> {
 
 			double distanceDesired = offsetX * offsetX + offsetY * offsetY + offsetZ * offsetZ;
 
-			distanceDesired = Mth.sqrt(distanceDesired);
+			distanceDesired = Mth.sqrt((float) distanceDesired);
 
 			// add velocity
 			double velX = offsetX / distanceDesired * 0.05D;
